Remove API MessagePackWriter.Write(sbyte[])
