+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Formatters/StandardClassLibraryFormatter.cs
@@ -9,6 +9,11 @@ using System.Globalization;
 using System.Text;
 using MessagePack.Internal;
 
+#if HARDWARE_INTRINSICS_X86
+using System.Runtime.Intrinsics;
+using System.Runtime.Intrinsics.X86;
+#endif
+
 #pragma warning disable SA1649 // File name should match first type name
 
 namespace MessagePack.Formatters
@@ -43,9 +48,342 @@ namespace MessagePack.Formatters
         {
         }
 
-        public void Serialize(ref MessagePackWriter writer, sbyte[] value, MessagePackSerializerOptions options)
+        public unsafe void Serialize(ref MessagePackWriter writer, sbyte[] value, MessagePackSerializerOptions options)
         {
-            writer.Write(value);
+            if (value == null)
+            {
+                writer.WriteNil();
+                return;
+            }
+
+            var inputLength = value.Length;
+            writer.WriteArrayHeader(inputLength);
+            if (inputLength == 0)
+            {
+                return;
+            }
+
+            fixed (sbyte* pSource = &value[0])
+            {
+                var inputEnd = pSource + inputLength;
+                var inputIterator = pSource;
+
+#if HARDWARE_INTRINSICS_X86
+                const int Stride = 16;
+                if (Popcnt.IsSupported && inputLength >= Stride)
+                {
+                    {
+                        // Make InputIterator Aligned
+                        var offset = new UIntPtr(inputIterator).ToUInt64();
+                        offset -= (offset >> 4) << 4;
+                        offset = (16 - offset) & 0xfUL;
+                        inputLength -= (int)offset;
+                        var offsetEnd = inputIterator + offset;
+                        while (inputIterator != offsetEnd)
+                        {
+                            writer.Write(*inputIterator++);
+                        }
+                    }
+
+                    ReadOnlySpan<byte> shuffle = new byte[]
+                    {
+                        0, 1, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 6, 7, 128, 128, 128,
+                        0, 1, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 6, 7, 128, 128, 128,
+                        0, 1, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 7, 128, 128,
+                        0, 1, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 7, 128, 128,
+                        0, 1, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 7, 128, 128,
+                        0, 1, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 7, 128,
+                        0, 1, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 6, 128, 7, 128, 128,
+                        0, 1, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 6, 128, 7, 128, 128,
+                        0, 1, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 6, 128, 7, 128,
+                        0, 1, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 1, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 5, 128, 6, 128, 7, 128,
+                        0, 1, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        0, 128, 1, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 1, 128, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 128, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 4, 128, 5, 128, 6, 128, 7, 128,
+                        0, 1, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128, 128,
+                        128, 0, 1, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        0, 128, 1, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 128, 1, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        128, 0, 128, 1, 128, 2, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128,
+                        0, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128, 128,
+                        128, 0, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        128, 0, 128, 1, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128,
+                        0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128, 128,
+                        128, 0, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128,
+                        0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7, 128,
+                        128, 0, 128, 1, 128, 2, 128, 3, 128, 4, 128, 5, 128, 6, 128, 7,
+                    };
+                    fixed (byte* pShuffle = shuffle)
+                    {
+                        var simdEnd = inputIterator + ((inputLength >> 4) << 4);
+                        while (inputIterator != simdEnd)
+                        {
+                            var current = Sse2.LoadAlignedVector128(inputIterator);
+                            var moveMask = unchecked((uint)Sse2.MoveMask(Sse2.CompareGreaterThan(Vector128.Create((sbyte)-32), current)));
+                            var count = unchecked((int)Popcnt.PopCount(moveMask)) + Stride;
+                            var destination = writer.GetSpan(count);
+                            fixed (byte* pDestination = &destination[0])
+                            {
+                                var moveMaskLower = moveMask & 0xff;
+                                var countLower = unchecked((int)Popcnt.PopCount(moveMaskLower)) + (Stride >> 1);
+                                var shuffleCurrentLower = Sse2.LoadVector128(pShuffle + (moveMaskLower << 4));
+                                var maskMoveMaskLower = Sse2.ShiftRightLogical128BitLane(Vector128.Create(byte.MaxValue), (byte)(16 - countLower));
+                                var shuffledLower = Ssse3.Shuffle(current.AsByte(), shuffleCurrentLower);
+                                var blendedLower = Sse41.BlendVariable(shuffledLower, Vector128.Create((byte)0xd0), shuffleCurrentLower);
+                                Sse2.MaskMove(blendedLower, maskMoveMaskLower, pDestination);
+
+                                current = Sse2.ShiftRightLogical128BitLane(current, 8);
+                                var moveMaskHigher = moveMask >> 8;
+                                var shuffleCurrentHigher = Sse2.LoadVector128(pShuffle + (moveMaskHigher << 4));
+                                var maskMoveMaskHigher = Sse2.ShiftRightLogical128BitLane(Vector128.Create(byte.MaxValue), (byte)(16 + countLower - count));
+                                var shuffledHigher = Ssse3.Shuffle(current.AsByte(), shuffleCurrentHigher);
+                                var blendedHigher = Sse41.BlendVariable(shuffledHigher, Vector128.Create((byte)0xd0), shuffleCurrentHigher);
+                                Sse2.MaskMove(blendedHigher, maskMoveMaskHigher, pDestination + countLower);
+                            }
+
+                            writer.Advance(count);
+                            inputIterator += Stride;
+                        }
+                    }
+                }
+#endif
+
+                while (inputIterator != inputEnd)
+                {
+                    writer.Write(*inputIterator++);
+                }
+            }
         }
 
         public sbyte[] Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
